Removed duplicate code
